use a symbolic link for the C++ code
Prior to this change users needed to source the file 'build_package.R'
which copied files from the general library source into the inst/include
directory.  This change uses a symbolic link so that this file and that
build step is no longer needed.

NOTE: If you are working on Linux of Mac the you should have not problem
cloning and working with this repo.  If, however, you are working on a
Windows machine there are a few additional steps you will need to take.

1. Windows Vista or newer with NTFS file system, not FAT.
2. You need administrator rights or at least `SeCreateSymbolicLinkPrivilege`
   privilege
3. git bash version 2.10.2 or later.  It will be helpful to install with
   `core.symlinks` option turned on.

In the git bash shell clone the repo.  (The example below uses SSH, change the
URL as needed for https.)

    git clone -c core.symlinks=true git@github.com:BayesPulse/libpulsatile