Update testing code and make functions inline
Previous C++ testing code updated so that it will
now run using testthat(). In order to make that
happen, several functions had to be made inline. At this time only ~218/225 of the tests pass, but
I have yet to determine if that is from errors in the functions or errors in the tests.