Handle patients with differing observation counts
Updates fit_pop() and several other functions
to enable them to work with patients with different
numbers of observations. As such, the main data
arg for fit_pop() is no longer a matrix, but now
a list of vectors. Similarly, time is now a list of vectors. Return data is now a list of data.frames